requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
shap
mlflow
matplotlib
seaborn
joblib
openai 
langchain 
streamlit
pillow